Move cdm-misc tests out of cdm-test.
