+++ cdm-test/build.gradle
@@ -12,7 +12,7 @@ dependencies {
     testCompile project(":cdm:cdm-core")
     testCompile project(":cdm:cdm-image")
     testCompile project(":cdm:cdm-radial")
-    testCompile project(":cdm:cdm-misc")
+    // testCompile project(":cdm:cdm-misc")
     testCompile project(":cdm:cdm-s3")
     testCompile project(":bufr")
     testCompile project(":grib")
